Add files via upload
I created a portfolio website that showcases my skills and professional profile. The website features a navigation menu for easy access to different sections like Home, About, Skills, Portfolio, and Contact. The left section of the site includes social media icons with animations, an introduction with my name, my profession as a Software Engineer, and a brief description of my skills and passion. The right section has a modern design with animated circular elements, adding a dynamic visual appeal to the website. The overall design uses animations, gradients, and hover effects to create an interactive and visually engaging experience.